+++ pycondor/tests/test_dagman.py
@@ -19,9 +19,7 @@ def test_job_dag_submit_file_same(tmpdir):
     # same whether built from a Dagman or not. See issue #38.
 
     example_script = os.path.join('examples/savelist.py')
-
     submit_dir = str(tmpdir.mkdir('submit'))
-
     # Build Job object that will be built outside of a Dagman
     job_outside_dag = pycondor.Job('job_outside_dag', example_script,
                                    submit=submit_dir, queue=5)
@@ -37,3 +35,33 @@ def test_job_dag_submit_file_same(tmpdir):
     # Check that the contents of the two Job submit files are the same
     assert filecmp.cmp(job_outside_dag.submit_file, job_inside_dag.submit_file,
                        shallow=False)
+
+
+def test_job_arg_name_files(tmpdir):
+    # Test to check that when a named argument is added to a Job, and the Job
+    # is built with fancyname=True, the Job submit file and the
+    # error/log/output files for the argument start with the same index.
+    # E.g. job_(date)_01.submit, job_(date)_01.error, etc.
+    # Regression test for issue #47
+    example_script = os.path.join('examples/savelist.py')
+    submit_dir = str(tmpdir.mkdir('submit'))
+
+    job = pycondor.Job('testjob', example_script, submit=submit_dir)
+    job.add_arg('arg', name='argname')
+    dagman = pycondor.Dagman('testdagman', submit=submit_dir)
+    dagman.add_job(job)
+    dagman.build(fancyname=True)
+
+    with open(dagman.submit_file, 'r') as dagman_submit_file:
+        dagman_submit_lines = dagman_submit_file.readlines()
+
+    # Get root of the dagman submit file (submit file basename without .submit)
+    submit_file_line = dagman_submit_lines[0]
+    submit_file_basename = submit_file_line.split('/')[-1].rstrip()
+    submit_file_root = os.path.splitext(submit_file_basename)[0]
+    # Get job_name variable (used to built error/log/output file basenames)
+    jobname_line = dagman_submit_lines[2]
+    jobname = jobname_line.split('"')[-2]
+    other_file_root = '_'.join(jobname.split('_')[:-1])
+
+    assert submit_file_root == other_file_root